boilerplate/working_python/requirements.txt
websockets==10.1
python-dotenv==0.19.0
